services/agent-service/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
google-genai
httpx
python-dotenv
